backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
pydantic
alembic
python-dotenv
